+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enhancements
   - #382: Support for cabal2nix generated expressions (avoiding IFD for local packages)
+  - Add `extraCabal2nixOptions` and `cabalFlags` to `packages.*`. The former passes custom options to `cabal2nix`. The later does the same, but formatting the flag options appropriately.
   - Support `meta.description` in flake apps. Requires newer version of flake-parts.
   - `settings` module:
     - #384: Add `stan`