+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Support `meta.description` in flake apps. Requires newer version of flake-parts.
   - `settings` module:
     - #384: Add `stan`
+    - #419: Add `installIntermediates` and `separateIntermediatesOutput`
 
 ## 0.5.0 (Jun 24, 2024)