+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - `settings` module:
     - #384: Add `stan`
     - #419, #421: Add `installIntermediates`, `previousIntermediates` and `separateIntermediatesOutput`
+    - #427: Add `generateOptparseApplicativeCompletions`
 
 ## 0.5.0 (Jun 24, 2024)