+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - #419, #421: Add `installIntermediates`, `previousIntermediates` and `separateIntermediatesOutput`
     - #427: Add `generateOptparseApplicativeCompletions`
     - #455: Add `drvAttrs` option for setting arbitrary derivation attributes
+    - #459: Apply `custom` ahead of other settings.
 
 ## 0.5.0 (Jun 24, 2024)